rename kh-tools to sencha
