+++ docs/index.rst
@@ -1,4 +1,4 @@
-.. khtools documentation master file, created by
+.. sencha documentation master file, created by
    sphinx-quickstart on Tue Jul  9 22:26:36 2013.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
