+++ sencha/compare_kmer_content.py
@@ -14,7 +14,7 @@ from sourmash.logging import notify
 
 # Divergence time estimates in millions of years
 # from http://www.timetree.org/ on 2019-08-26
-from khtools.sequence_encodings import amino_keto_ize, \
+from sencha.sequence_encodings import amino_keto_ize, \
     weak_strong_ize, purine_pyrimidize, encode_peptide
 
 MOLECULES_TO_COMPARE = 'peptide20', 'hsdm17', 'sdm12', 'aa9', 'botvinnik8', \
