+++ sencha/downsample_hashes.py
