+++ sencha/extract_metadata.py
