+++ sencha/knn.py
