+++ sencha/sequence_encodings.py
