+++ sencha/translate_single_seq.py
