+++ tests/test_ensembl.py
@@ -69,7 +69,7 @@ def ensembl_mock(url, request):
 
     try:
         filename = os.path.join(pytest.config.rootdir,
-                                f'khtools/tests/data/{ensembl_id}.pkl')
+                                f'sencha/tests/data/{ensembl_id}.pkl')
         with open(filename, 'rb') as testing_file:
             testing_data = pickle.load(testing_file)
     except FileNotFoundError:
@@ -91,7 +91,7 @@ def ensembl_mock(url, request):
 
 @pytest.mark.skip
 def test_lookup(ensembl_protein_id):
-    from khtools.ensembl import lookup
+    from sencha.ensembl import lookup
 
     with HTTMock(ensembl_mock):
         test = lookup(ensembl_protein_id)
@@ -102,7 +102,7 @@ def test_lookup(ensembl_protein_id):
 
 @pytest.mark.skip
 def test_lookup_expand_true(ensembl_transcript_id):
-    from khtools.ensembl import lookup
+    from sencha.ensembl import lookup
 
     with HTTMock(ensembl_mock):
         test = lookup(ensembl_transcript_id, expand=True)
@@ -146,7 +146,7 @@ def test_lookup_expand_true(ensembl_transcript_id):
 
 @pytest.mark.skip
 def test_get_orthologues(ensembl_gene_id):
-    from khtools.ensembl import get_orthologues
+    from sencha.ensembl import get_orthologues
 
     with HTTMock(ensembl_mock):
         test = get_orthologues(ensembl_gene_id, 'mouse')
