+++ tests/test_os_utils.py
@@ -18,7 +18,7 @@ def folder():
 
 
 def test_sanitize_path():
-    from khtools.os_utils import sanitize_path
+    from sencha.os_utils import sanitize_path
 
     test = sanitize_path('.')
     true = os.path.abspath('.')
@@ -26,14 +26,14 @@ def test_sanitize_path():
 
 
 def test_maybe_add_slash(folder):
-    from khtools.os_utils import maybe_add_slash
+    from sencha.os_utils import maybe_add_slash
 
     test = maybe_add_slash(folder)
     assert test == 'test-folder/'
 
 
 def test_get_stdout_from_command():
-    from khtools.os_utils import get_stdout_from_command
+    from sencha.os_utils import get_stdout_from_command
     command = ['echo', 'asdf']
     stdout = get_stdout_from_command(command)
     assert stdout == ['asdf']
@@ -42,7 +42,7 @@ def test_get_stdout_from_command():
 @pytest.mark.skipif(not sys.platform.startswith("darwin"),
                     reason="computers are hard")
 def test_get_stdout_stderr_from_command():
-    from khtools.os_utils import get_stdout_stderr_from_command
+    from sencha.os_utils import get_stdout_stderr_from_command
 
     command = ['sed', 'asdf']
     stdout, stderr = get_stdout_stderr_from_command(command)
